net, istio: Change ReservedPorts's signature
When Istio is used alongside the passt binding plugin, its sidecar
specifies the list of Istio's reserved ports in the domain XML.
The domain XML uses the `uint` type in order to specify the Start field
on the `InterfacePortForwardRange` struct.

Currently, `istio.ReservedPorts` returns a slice of integers, which is
converted into `uint`.

Adjust `istio.ReservedPorts` signature to return a slice of unsigned
integers, in order to fix a potential integer overflow pointed out by
the linter.

+++ cmd/sidecars/network-passt-binding/domain/configurator.go
@@ -200,7 +200,7 @@ func (p PasstNetworkConfigurator) generatePortForward() []domainschema.Interface
 
 	if p.options.IstioProxyInjectionEnabled {
 		for _, port := range istio.ReservedPorts() {
-			tcpPortsRange = append(tcpPortsRange, domainschema.InterfacePortForwardRange{Start: uint(port), Exclude: "yes"})
+			tcpPortsRange = append(tcpPortsRange, domainschema.InterfacePortForwardRange{Start: port, Exclude: "yes"})
 		}
 	}
